Support extended pdbid (#950)
* Support for extended PDBID PDB_nnnnNXXX

Created a new class PDBId, and removed old deprecated methods.
Most of the changes are in Structure, StructureImpl, and PDBFileParser
classes.

* Some deprecated functions removal

* Fix typo

* Targeted most of Spencer's comments

* Fixed a bug in PDBId.equals() method

* Changing PDBID in SCOP record to lower case

to match the expected String

* public static PDBId field XXXX

* Behavior is now an enum

* Expose Pattern objects instead of Strings

* ID stored internally in a reduced format (without initial "PDB_" prefix)

Only the ID proper is stored (in upper case).
e.g. for 1abc we store 00001ABC.

This format is the trade off between storing short/extended format and
all extended format.
storing the ID proper allows direct comparison and fast ordering of
PDBId objects with uniform names in the format \d{4}[1-9][0-9A-Z]{3}.

* Enhancements and bug fix

* Added BDBId Unit test

* Added @Deprecated codeTag

* Adding more TODOs

* more replacement of deprecated methods

* Removed PDBIdException

I replaced the PDBException with IllegalArgumentException or
StructureException:
- Constant literal extraction
- Inlined short private methods
- Renamed isShortPDBId() and isExtendedPDBID() to isValidShortPDBID()
and isValidExtendedPDBID()

Additional modifications:
- Organized imports

* review and untouch regular expressions

* Un-deprecate, copy javadoc, and remove @author on methods

* Remove NullPointerException

I will write the full class javadoc later

* Bug fix

* clean commented out code

* Update CHANGELOG

Mentioned that getReducedStructure() is removed from StructureTools

* Fix unit test

I accident.y found that the test case is not written as it should be.

* PdbPair does not accept null

It also considers an object and its reverse equal

* Addressing reviewer's comments

* Change capitalization state

- Class name: changed into PdbId
- methods name: changed into getPdbId(), setPdbId(PdbId)
- field name: pdbId

* Fix probable NPE + keep consistent PdbId method naming convention

* Addressed some of the reviewer's comments.

* Use JUnit 5

* PdbId class Documentation

* JavaDoc style update

* Minor updates

* Reverting a wrong optimization

It introduced a new bug. Reverting it.

* Adding BioJava development code

* Javadoc

* XXXX PdbId objects are not equal unless they are the same object

* No more XXXX. ANY malformed PdbId gracefully set to null

* Addressing Reviewer's comments

* Removing TODOs